# Conflicts:
#	AddOns/Agenda/server/TKCachedSkedgoifier.swift
#	AddOns/Agenda/server/TKSkedgoifier.h
#	AddOns/Agenda/server/TKSkedgoifier.m
#	AddOns/Agenda/server/TKSkedgoifier.swift
#	AddOns/Agenda/workers/TKAgendaFaker.swift

+++ Classes/server/TKBuzzRouter.m
@@ -154,7 +154,7 @@
 {
   if (!identifier) {
     NSUInteger hash = [[url absoluteString] hash];
-    identifier = [NSString stringWithFormat:@"%lu", hash];
+    identifier = [NSString stringWithFormat:@"%lu", (unsigned long) hash];
   }
   
   TKJSONCacheDirectory directory = TKJSONCacheDirectoryDocuments;
